send more data from invoker back to controller

+++ core/invoker/src/main/scala/org/apache/openwhisk/core/containerpool/UpdatablePriorityQueue.scala
@@ -6,8 +6,8 @@ package org.apache.openwhisk.core.containerpool
 class TrackedAction() {
     var lastcalled : Double = 0.0;
     var invocations : Long = 0;
-    var coldTime : Long = 0;
-    var warmTime : Long = 0;
+    var coldTime : Double = 0;
+    var warmTime : Double = 0;
     var memory : Double = 0.0;
     var active : Long = 0;
